server/requirements.txt
fastapi
uvicorn
python-dotenv
pydantic-settings
tavily-python
trafilatura
sentence-transformers
numpy
google-generativeai